requirements_post_install.txt
absl-py==0.15.0
astunparse==1.6.3
cached-property==1.5.2
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
flatbuffers==1.12
gast==0.4.0
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
h5py==3.1.0
idna==3.3
importlib-metadata==4.11.0
keras-nightly==2.6.0.dev2021062500
Keras-Preprocessing==1.1.2
libclang==11.1.0
Markdown==3.3.6
numpy==1.21.5
oauthlib==3.2.0
opt-einsum==3.3.0
portpicker==1.5.0
protobuf==3.19.4
psutil==5.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.15.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow @ file:///tmp/tensorflow_pkg/tensorflow-2.7.0-cp37-cp37m-linux_x86_64.whl
tensorflow-estimator==2.5.0
termcolor==1.1.0
typing-extensions==3.7.4.3
urllib3==1.26.8
Werkzeug==2.0.3
wrapt==1.12.1
zipp==3.7.0